% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTranslation.R
\name{checkXlsformTranslations}
\alias{checkXlsformTranslations}
\title{Check Translations in an XLSForm Excel File}
\usage{
checkXlsformTranslations(xlsformPath)
}
\arguments{
\item{xlsformPath}{Character. Path to the XLSForm Excel file.}
}
\value{
Logical. TRUE if no missing translations are found, FALSE otherwise.
}
\description{
This function reads the 'survey' and 'choices' sheets of an XLSForm Excel file
and checks for missing translations across language columns (e.g., label, hint, etc.).
}
\examples{
# checkXlsformTranslations("data/treeCensusForm.xlsx")
}
