% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLangForm.R
\name{createMultiLangForm}
\alias{createMultiLangForm}
\title{Create a Multi-Language ODK Form}
\usage{
createMultiLangForm(baseForm, translationFile, outputFile)
}
\arguments{
\item{baseForm}{Character. Path to the base XML form file.}

\item{translationFile}{Character. Path to the CSV file containing translations.}

\item{outputFile}{Character. Path where the generated XML form should be saved.}
}
\value{
Character. The path to the generated output file.
}
\description{
This function takes a base XML form and a translation CSV file to generate
a multi-language ODK form with dynamic language selection.
}
\examples{
# createMultiLangForm("base_form.xml", "translations.csv", "multi_lang_form.xml")
}
