% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInstancesXML.R
\name{.lastSegment}
\alias{.lastSegment}
\title{Return the last segment of a slash-separated path}
\usage{
.lastSegment(path)
}
\arguments{
\item{path}{A slash-separated path.}
}
\value{
The last element name.
}
\description{
Return the last segment of a slash-separated path
}
\keyword{internal}
