% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapsvg.R
\name{svgMap}
\alias{svgMap}
\title{Graphic of trees maped}
\usage{
svgMap(
  mapData,
  subPlotCode = "A00",
  svgSave = TRUE,
  wd2save = file.path(getwd(), subPlotCode),
  dx = "dx",
  dy = "dy",
  tag = "tag",
  dbh = "dbh",
  status = "status",
  mapSize = c(13, 13),
  vpSize = c(0.9, 0.9),
  fontSize = 12,
  diagonal = FALSE
)
}
\arguments{
\item{mapData}{list with data for each subquadrat with a buffer zone.}

\item{subPlotCode}{Character string that indetified subplot code.}

\item{svgSave}{logical, if TRUE the device is exported to a svg file.}

\item{wd2save}{character string indicating the directory to save the svg file}

\item{dx}{name of the variável in censoData that contains the subplot x coordinate of the trees.}

\item{dy}{name of the variável in censoData that contains the subplot y coordenate of the trees.}

\item{tag}{character string with the name of the variável in censoData that contains the tag code that identifies trees.}

\item{dbh}{character string with the name of the variable diameter at breast high.}

\item{status}{character string with the name of the variable with the status of the tree ("A" = alive, "D" = dead).}

\item{mapSize}{size of the screen device window in inchs.}

\item{vpSize}{proportion of plot area in 'mapSize'.}

\item{fontSize}{numeric defining the font size.}

\item{diagonal}{logical, if TRUE the diagonals will be ploted.}
}
\value{
'svgGrid' returns figure device and export a svg file.
}
\description{
Map of trees from permanent plot based on cartesian coordinates
}
\examples{
\dontrun{
 svgMap(dataplot, svgSave = FALSE)
}
}
\seealso{
\code{\link[gridSVG]{gridSVG}}
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
