% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapsvg.R
\name{auditsvg}
\alias{auditsvg}
\title{Map of trees from permanent plot based on cartesian coordinates}
\usage{
auditsvg(
  audit,
  quad = "A00",
  save.svg = TRUE,
  wd = getwd(),
  dx = "new_dx2018",
  dy = "new_dy2018",
  tag = "num_tag",
  dap = "dap2018",
  error = "errorType",
  mapSize = c(13, 13)
)
}
\arguments{
\item{audit}{auditory data frame from permanent plot tree censo data.}

\item{quad}{a character string or factor representing some subunit of the permanent plot.}

\item{save.svg}{logical true to save a svg file}

\item{wd}{character string indicating directory towhere the file will be saved.}

\item{dx}{between 0 to max.size . Tree mapping X coordinate.}

\item{dy}{between 0 to max.size . Tree mapping Y coordinate.}
}
\value{
'mapsvg' returns tree mapped svg mapped with svg pattern ids.
}
\description{
Map of trees from permanent plot based on cartesian coordinates
}
\examples{

\dontrun{
dataplot <- data.frame(dx = runif(100, 0,20), dy = runif(100,0,20), quad = rep(paste("quad", 0:1, sep="_"), each=50), dap =rnbinom(100,10,0.5), status="A", tag = 1:100)
auditsvg(dataplot, save.svg =FALSE)
}

}
\seealso{
\code{\link{gridSVG}}
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
