% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushToODK.R
\name{mediaToODK}
\alias{mediaToODK}
\title{Copy media files to ODK Collect on Android via ADB}
\usage{
mediaToODK(localMediaDir, delPrevMaps = TRUE, formName = "odkTreeCensus")
}
\arguments{
\item{localMediaDir}{Path to the local media directory.}

\item{delPrevMaps}{Logical. If TRUE, removes old map and grid SVG files on the device.}

\item{formName}{Name of the form (defaults to "odkTreeCensus").}
}
\value{
Returns a string indicating success and the number of files copied.
}
\description{
Copy media files to ODK Collect on Android via ADB
}
