% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushToODK.R
\name{pushToODKCollect}
\alias{pushToODKCollect}
\title{Copy local files or directories to ODK Collect on Android via ADB}
\usage{
pushToODKCollect(
  localDir = NULL,
  projectUUID = NULL,
  media = FALSE,
  formName = "treeCensusForm"
)
}
\arguments{
\item{localDir}{Optional. Path to the local directory containing the form and media. If NULL, defaults to the package's "odk" folder.}

\item{projectUUID}{Optional. The ODK project UUID. If NULL, the function will try to list and use the first project found.}

\item{media}{Logical. If TRUE, copies the "*-media" directory of the form.}

\item{formName}{Optional. The name of the form (without extension). Defaults to "odkTreeCensusForm".}
}
\value{
Returns TRUE if the copy was successful.
}
\description{
Copy local files or directories to ODK Collect on Android via ADB
}
