% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushToODK.R
\name{pushToODKCollect}
\alias{pushToODKCollect}
\title{Copy local files or directories to ODK Collect on Android via ADB}
\usage{
pushToODKCollect(
  localFiles,
  projectUUID = NULL,
  media = FALSE,
  formName = NULL
)
}
\arguments{
\item{localFiles}{Vector of local file or directory paths to be copied.}

\item{projectUUID}{Optional. The ODK project UUID. If NULL, the function will try to list and use the first project found.}

\item{media}{Logical. If TRUE, copies the files to the "*-media" directory of the form.}

\item{formName}{Optional. The name of the form (without extension) to define the media folder. If NULL and media is TRUE, the function will try to infer it from the local files.}
}
\value{
Returns TRUE if the copy was successful for all files/directories.
}
\description{
Copy local files or directories to ODK Collect on Android via ADB
}
