% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapsvg.R
\name{svgGrid}
\alias{svgGrid}
\title{Grid for mapping}
\usage{
svgGrid(
  censoData,
  subPlotCode = "A00",
  subqSize = 5,
  gridSize = 0.2,
  svgSave = TRUE,
  wd2save = file.path(getwd(), subPlotCode),
  dx = "dx",
  dy = "dy",
  tag = "tag",
  dbhcm = "dbhcm",
  status = "status",
  subquad = "subquad",
  mapSize = c(13, 13),
  vpSize = c(0.8, 0.8),
  fontSize = 12,
  diagonal = FALSE
)
}
\arguments{
\item{censoData}{data frame from permanent plot tree censo cartezian data.}

\item{subPlotCode}{Character string that indetified subplot code.}

\item{gridSize}{numeric vector. Size of the grid cell for mapping.}

\item{svgSave}{logical, if TRUE the device is exported to a svg file.}

\item{wd2save}{character string indicating the directory to save the svg fi}

\item{dx}{name of the variável in censoData that contains the subplot x coord}

\item{dy}{name of the variável in censoData that contains the subplot y coordenate of the trees.}

\item{status}{the size of the subquadrat to map in each graphic.}

\item{subquad}{a character string of the name of the subquad variable, representing some subunit of the subplot.}

\item{mapSize}{size of the figure in inchs.}

\item{diagonal}{logical, if TRUE the diagonals will be plot.}
}
\value{
'svgGrid' returns figure device and export a svg file.
}
\description{
A figure to select mapping position
}
\examples{
\dontrun{
 svgGrid(dataplot, svgSave = FALSE)
}

}
\seealso{
\code{\link{gridSVG}}
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
