% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlForm.R
\name{xmlForm}
\alias{xmlForm}
\title{Convert XLSForm to ODK XML using pyxform}
\usage{
xmlForm(xlsxPath, xmlPath = NULL)
}
\arguments{
\item{xlsxPath}{Character. Path to the input Excel (.xlsx) file.}

\item{xmlPath}{Character. Path where the output XML file should be saved.
If NULL, it will be saved in the same directory with a .xml extension.}
}
\value{
Character. The path to the generated XML file, invisibly.
}
\description{
This function converts an Excel XLSForm (.xlsx) into an ODK-compliant XML file
using the Python library \code{pyxform} bundled within the R package.
It also performs a basic integrity check on the generated XML file.
}
\examples{
\dontrun{
xmlForm("data/treeCensusForm.xlsx", "data/treeCensusForm.xml")
}
}
